Scripting/EntryExam/source/requirements.txt
Flask==1.0.2
Pillow==3.1.2
